requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==1.11.20
django-filter==1.0.2
django-heroku==0.3.1
django-import-export==1.2.0
djangorestframework==3.9.1
et-xmlfile==1.0.1
gunicorn==19.9.0
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.0
psycopg2==2.7.7
pytz==2018.9
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
